refactor: refactor CardsSection
