+++ products_application/src/components/Pages/MainPage/CardsSection/Card/Card.module.scss
