+++ products_application/src/components/Pages/MainPage/CardsSection/CardsSection.tsx
@@ -1,0 +1,22 @@
+import { IProduct } from '../../../../types';
+import { Card } from './Card/Card';
+import styles from './CardsSection.module.scss';
+
+export function CardsSection(props: {
+  products: IProduct[];
+}): React.ReactElement {
+  const { products } = props;
+  return (
+    <section className={styles.cards_section}>
+      {products.length > 0 ? (
+        products.map((product: IProduct) => {
+          return <Card product={product} key={product.id} />;
+        })
+      ) : (
+        <p className={styles.message}>
+          Sorry, nothing found. Please, try again!
+        </p>
+      )}
+    </section>
+  );
+}
