+++ products_application/src/components/Pages/MainPage/MainPage.tsx
@@ -2,11 +2,10 @@ import { useState } from 'react';
 import { Loader } from '../../Loader/Loader';
 import styles from './MainPage.module.scss';
 import { SearchSection } from './SearchSection/SearchSection';
-import { CardsSection } from './SearchSection/CardsSection/CardsSection';
+import { CardsSection } from './CardsSection/CardsSection';
 
 export function MainPage(): React.ReactElement {
   const [isLoading, setIsLoading] = useState(false);
-  const [isNothingFound, setIsNothingFound] = useState(false);
   const [products, setProducts] = useState([
     {
       id: 1,
@@ -26,7 +25,6 @@ export function MainPage(): React.ReactElement {
 
   const handleSubmitButton = (): void => {
     setIsLoading(true);
-    setIsNothingFound(false);
     setProducts([
       {
         id: 1,
@@ -51,11 +49,6 @@ export function MainPage(): React.ReactElement {
       <main className={styles.main}>
         <SearchSection onSubmit={handleSubmitButton} />
         {isLoading ? <Loader /> : <CardsSection products={products} />}
-        {isNothingFound && (
-          <p className={styles.message}>
-            Sorry, nothing found. Please, try again!
-          </p>
-        )}
       </main>
       <footer className={styles.footer}>Footer</footer>
     </>
