--- products_application/src/components/Pages/MainPage/SearchSection/CardsSection/CardsSection.module.scss
@@ -1,8 +1,0 @@
-.cards_section {
-    max-width: 70%;
-    position: relative;
-    justify-content: space-between;
-    display: flex;
-    margin: 10px auto;
-    flex-wrap: wrap;
-}
