+++ products_application/src/components/Pages/MainPage/SearchSection/SearchBlock/SearchBlock.module.scss
@@ -31,16 +31,3 @@
     padding: 5px 5px;
     border-radius: 5px;
 }
-
-.message {
-    position: absolute;
-    top: 45vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-weight: bold;
-    font-size: 30px;
-    font-family: monospace;
-    width: 100%;
-    margin: 0 auto;
-}
